+++ src/utils/index.ts
@@ -1,4 +1,1 @@
-export * from './isPropertyRequired';
 export * from './printName';
-export * from './isParentNode';
-export * from './schemaUtils';
