--- src/utils/isParentNode.ts
@@ -1,1 +1,0 @@
-import { isRegularNode, SchemaNode } from '@stoplight/json-schema-tree';
